Add more domain checks

+++ test/checks/compression.ts
@@ -1,6 +1,6 @@
 // Check that pages are compressed better than by gzip
 
-import { fail, pass } from './utils.ts'
+import { fail, pass, short } from './utils.ts'
 
 export async function checkCompression(pages: string[]): Promise<boolean> {
   let results = []
@@ -12,12 +12,14 @@ export async function checkCompression(pages: string[]): Promise<boolean> {
     let used = response.headers.get('content-encoding')
     if (used !== 'br' && used !== 'zstd') {
       results.push(
-        fail(`${url} has ${used ?? 'no'} compression instead of zstd`)
+        fail(`${short(url)} has ${used ?? 'no'} compression instead of zstd`)
       )
     } else if (!response.headers.get('vary')?.includes('Accept-Encoding')) {
-      results.push(fail(`${url} has no Vary: Accept-Encoding for caches`))
+      results.push(
+        fail(`${short(url)} has no Vary: Accept-Encoding for caches`)
+      )
     } else {
-      results.push(pass(`${url} is compressed by ${used}`))
+      results.push(pass(`${short(url)} is compressed by ${used}`))
     }
   }
   return results.every(Boolean)
